refactor: extract relay access-request wait flow into helper module
